+++ doc/whats-new.rst
@@ -34,6 +34,8 @@ New Features
 Bug fixes
 ~~~~~~~~~
 
+- Fix bug where reading a scalar value from a NetCDF file opened with the ``h5netcdf`` backend would raise a ``ValueError`` when ``decode_cf=True`` (:issue:`4471`, :pull:`4485`).
+  By `Gerrit Holl <https://github.com/gerritholl>`_.
 - Fix bug where datetime64 times are silently changed to incorrect values if they are outside the valid date range for ns precision when provided in some other units (:issue:`4427`, :pull:`4454`).
   By `Andrew Pauling <https://github.com/andrewpauling>`_
 - Fix silently overwriting the `engine` key when passing :py:func:`open_dataset` a file object